requirements.txt
asgiref==3.8.1
astroid==3.2.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dill==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
install==1.3.5
isort==5.13.2
Markdown==3.6
mccabe==0.7.0
msgpack==1.0.8
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pylint==3.2.4
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==6.4.post2
